Implementing Foundry testing file, fixed some bugs\typos and wrote some tests by @Perelyn-sama

+++ Maths/armstrongNumber/armstrongNumber.sol
@@ -1,15 +1,15 @@
 pragma solidity ^0.8.0;
 
 /** @title This is a Solidity implementation to Armstrong number.
-  * @notice An Armstrong number is number that is equal to the sum of
-  * cubes of its digits.
-  * @author Wenceslas Sanchez (wenceslas-sanchez)
-  */
+ * @notice An Armstrong number is a number that is equal to the sum of
+ * cubes of its digits.
+ * @author Wenceslas Sanchez (wenceslas-sanchez)
+ */
 contract ArmstrongNumber {
     /** @notice Return the digit numbers of an integer.
-      * @dev From https://ethereum.stackexchange.com/questions/6623/can-i-check-the-number-of-digits-in-solidity
-      */
-    function numDigits(int _n) public pure returns (uint) {
+     * @dev From https://ethereum.stackexchange.com/questions/6623/can-i-check-the-number-of-digits-in-solidity
+     */
+    function numDigits(int256 _n) public pure returns (uint256) {
         uint8 digits = 0;
         while (_n != 0) {
             _n /= 10;
@@ -20,21 +20,22 @@ contract ArmstrongNumber {
     }
 
     /** @notice Function to check whether a number is armstrong or not.
-      * @return `true` if the number is armstrong
-      */
-    function isArmstrongNumber(uint _n) public pure returns (bool) {
-        uint sum= 0;
-        uint saveNum= _n;
-        uint rightNum;
+     * @return `true` if the number is armstrong
+     */
+    function isArmstrongNumber(uint256 _n) public pure returns (bool) {
+        uint256 sum = 0;
+        uint256 saveNum = _n;
+        uint256 rightNum;
 
-        uint digits= numDigits(int(_n));
+        uint256 digits = numDigits(int256(_n));
 
         while (saveNum > 0) {
             rightNum = saveNum % 10;
-            sum += rightNum ** digits;
+            sum += rightNum**digits;
             saveNum /= 10;
         }
 
         return _n == sum;
     }
 }
+
